pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
authors = ["Crossbar.io Technologies GmbH", "Aki Mimoto <aki@zaber.com>"]
description = "Patched version to integrate into AD based network a bit easier. Intended for internal use."
packages = [
  { include = "crossbar" },
  { include = "nexus" },
]

[tool.poetry.dependencies]
python = "^3.7.2"
attrs = ">=17.2.0"
base58 = ">=2.1.0"
bcrypt = ">=3.1.6"
bitstring = ">=3.1.5"
cbor = ">=1.0.0"
click = ">=7.0"
colorama = ">=0.4.4"
constantly = ">=15.1.0"
cookiecutter = ">=1.6.0"
cryptography = ">=2.6.1"
docker = ">=3.5.0"
eth-account = ">=0.5.4"
eth-typing = ">=2.2.2"
Flask = ">=1.1.1"
h2 = ">=3.2.0,<4.0.0"
hexbytes = ">=0.2.1"
humanize = ">=0.5.1"
hyperframe = ">=5.2.0,<6.0.0"
idna = "<2.6,>=2.5"
importlib-resources = ">=4.1.1"
incremental = ">=17.5.0"
iso8601 = ">=0.1.12"
jinja2-highlight = ">=0.6.1"
Jinja2 = ">=2.10.1"
lmdb = ">=1.1.1"
MarkupSafe = ">=1.1.1"
mistune = ">=0.8.4"
morphys = ">=1.0"
netaddr = ">=0.8.0"
numpy = ">=1.20.0"
parsimonious = ">=0.8.1"
passlib = ">=1.7.1"
priority = ">=1.3.0,<2.0"
prompt-toolkit = ">=2.0.10"
psutil = ">=5.8.0"
py-eth-sig-utils = ">=0.4.0"
py-ubjson = ">=0.9.0"
pyasn1-modules = ">=0.2.4"
pyasn1 = ">=0.4.5"
pycryptodome = ">=3.7.1"
Pygments = ">=2.2.0"
PyNaCl = ">=1.1.2"
pyOpenSSL = ">=17.1.0"
PyQRCode = ">=1.2.1"
PyTrie = ">=0.3"
PyYAML = ">=4.2b4"
requests = ">=2.20.0"
sdnotify = ">=0.3.1"
service-identity = ">=17.0.0"
setproctitle = ">=1.1.10"
setuptools = ">=36.2.7"
six = ">=1.13.0"
stringcase = ">=1.2.0"
tabulate = ">=0.7.7"
treq = ">=20.4.1"
txtorcon = ">=20.0.0"
u-msgpack-python = ">=2.4.1"
urllib3 = ">=1.21.1,<1.27"
validate_email = ">=1.3"
vmprof = ">=0.4.15"
watchdog = ">=0.8.3"
Werkzeug = ">=0.14.1"
xbr = ">=21.2.1"
"zope.interface" = ">=4.4.2"
izaber = ">=2.20210919"
izaber-wamp = "^2.20210228"
bs4 = "^0.0.1"
Faker = "^8.13.2"
ldap3 = "^2.9.1"
pytz = "^2021.1"
schedule = "^1.1.0"
"ruamel.yaml" = "^0.17.16"
ruyaml = "^0.20.0"
zlmdb = {git = "https://github.com/crossbario/zlmdb.git", rev = "master"}
txaio = {git = "https://github.com/crossbario/txaio.git", rev = "master"}
py-cid = {git = "https://github.com/crossbario/py-cid.git", rev = "remove-dep-upper-limits"}
py-multihash = {git = "https://github.com/crossbario/py-multihash.git", rev = "remove-dep-version-limits"}
ipfshttpclient = "^0.7.0"
web3 = ">=5.13.1"
autobahn = {git = "https://github.com/crossbario/autobahn-python.git", rev = "master", extras = ["twisted", "encryption", "compress", "serialization", "scram", "xbr"]}
Twisted = {version = ">=20.3.0", extras = ["tls", "conch", "http2"]}
cfxdb = "^21.2.1"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]

crossbar = "crossbar:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
